+++ src/components/layouts/flipCard/FlipCard.tsx
@@ -1,61 +1,10 @@
 import { useState } from 'react';
 import { Box, Typography } from '@mui/material';
-import { styled } from '@mui/system';
-import theme from '../../../themes'; // Adjust the import according to your project structure
 import { QuestionInterface } from '../../../interfaces/question';
 import FlipCardFrontFace from './FlipCardFrontFace';
 import FlipCardBackFace from './FlipCardBackFace';
 import { Lesson } from '../../../interfaces/lessons';
-
-interface FlipCardInnerProps {
-	isFlipped: boolean;
-}
-
-const FlipCardContainer = styled(Box)({
-	position: 'relative',
-	display: 'flex',
-	width: '50vh',
-	height: '40vh',
-	perspective: '50rem',
-	margin: '0 auto 3rem auto',
-	cursor: 'pointer',
-});
-
-const FlipCardInner = styled(Box, {
-	shouldForwardProp: (prop) => prop !== 'isFlipped',
-})<FlipCardInnerProps>(({ isFlipped }) => ({
-	position: 'absolute',
-	width: '100%',
-	height: '100%',
-	transition: 'transform 0.6s',
-	transformStyle: 'preserve-3d',
-	transform: isFlipped ? 'rotateY(180deg)' : 'rotateY(0deg)',
-}));
-
-const FlipCardSide = styled(Box)({
-	position: 'absolute',
-	width: '100%',
-	height: '100%',
-	backfaceVisibility: 'hidden',
-	display: 'flex',
-	flexDirection: 'column',
-	justifyContent: 'center',
-	alignItems: 'center',
-	padding: '2rem',
-	color: 'white',
-	borderRadius: '0.5rem',
-	boxShadow: '0 2px 4px rgba(0, 0, 0, 0.1)',
-	overflow: 'hidden', // Ensure no overflow
-});
-
-const FlipCardFront = styled(FlipCardSide)({
-	backgroundColor: theme.bgColor?.greenPrimary,
-});
-
-const FlipCardBack = styled(FlipCardSide)({
-	backgroundColor: 'coral',
-	transform: 'rotateY(180deg)',
-});
+import FlipCardPreview from './FlipCardPreview';
 
 interface FlipCardProps {
 	newQuestion?: QuestionInterface;
@@ -84,13 +33,8 @@ const FlipCard = ({
 	fromLessonEditPage,
 	imageUrlAdminQuestions,
 }: FlipCardProps) => {
-	const [isFlipped, setIsFlipped] = useState<boolean>(false);
 	const [frontText, setFrontText] = useState<string>(question?.question || newQuestion?.question || '');
 	const [backText, setBackText] = useState<string>(question?.correctAnswer || newQuestion?.question || '');
-
-	const handleClick = () => {
-		setIsFlipped(!isFlipped);
-	};
 
 	return (
 		<Box sx={{ width: '100%' }}>
@@ -123,53 +67,15 @@ const FlipCard = ({
 					Preview
 				</Typography>
 			</Box>
-			<FlipCardContainer>
-				<FlipCardInner isFlipped={isFlipped} onClick={handleClick}>
-					<FlipCardFront>
-						{(question?.imageUrl || newQuestion?.imageUrl) && (
-							<img
-								src={setNewQuestion ? newQuestion?.imageUrl : fromLessonEditPage ? question?.imageUrl : imageUrlAdminQuestions}
-								alt='img'
-								style={{
-									width: '100%',
-									height: question?.question ? '70%' : '90%',
-									objectFit: 'contain',
-									borderRadius: '0.5rem 0.5rem 0 0',
-								}}
-							/>
-						)}
-						<Box
-							sx={{
-								marginTop: question?.imageUrl ? '1rem' : 0,
-								textAlign: 'center',
-								padding: '0 1rem',
-								maxHeight: question?.imageUrl ? 'calc(70% - 1rem)' : '100%',
-								overflow: 'auto',
-							}}>
-							<Typography
-								variant={question?.imageUrl ? 'body2' : 'body1'}
-								sx={{
-									whiteSpace: 'pre-wrap',
-									wordWrap: 'break-word',
-									textOverflow: 'ellipsis',
-									color: theme.textColor?.common.main,
-								}}>
-								{frontText}
-							</Typography>
-						</Box>
-					</FlipCardFront>
-
-					<FlipCardBack>
-						<Typography
-							variant='body1'
-							sx={{
-								color: theme.textColor?.common.main,
-							}}>
-							{backText}
-						</Typography>
-					</FlipCardBack>
-				</FlipCardInner>
-			</FlipCardContainer>
+			<FlipCardPreview
+				question={question}
+				newQuestion={newQuestion}
+				fromLessonEditPage={fromLessonEditPage}
+				imageUrlAdminQuestions={imageUrlAdminQuestions}
+				frontText={frontText}
+				backText={backText}
+				setNewQuestion={setNewQuestion}
+			/>
 		</Box>
 	);
 };
